Create AndroidOS.java
Rather than running on its own, Question.java will play a small part in your final Android game. To see an example of how this works, navigate to the AndroidOS.java file.
The AndroidOS class serves as a representation of the Android application layer (the part of Android that engages the user and the screen directly).